+++ client/src/components/AdminBoard.tsx
@@ -1,32 +1,25 @@
-import { Component } from "react";
+import { useState, useEffect } from "react";
 
 import UserService from "../services/user.service";
 import EventBus from "../common/EventBus";
 
-type Props = {};
 
 type State = {
   content: string;
 }
 
-export default class BoardAdmin extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
+const BoardAdmin =() => {
+  const [state, setState] = useState<State>({content: ""})
 
-    this.state = {
-      content: ""
-    };
-  }
-
-  componentDidMount() {
-    UserService.getModeratorBoard().then(
+  useEffect(() => {
+    UserService.getAdminBoard().then(
       response => {
-        this.setState({
+        setState({
           content: response.data
         });
       },
       error => {
-        this.setState({
+        setState({
           content:
             (error.response &&
               error.response.data &&
@@ -40,15 +33,15 @@ export default class BoardAdmin extends Component<Props, State> {
         }
       }
     );
-  }
+  }, []) 
 
-  render() {
+
     return (
       <div className="container">
         <header className="jumbotron">
-          <h3>{this.state.content}</h3>
+          <h3>{state?.content}</h3>
         </header>
       </div>
     );
-  }
 }
+export default BoardAdmin;
