+++ client/src/design/react-typescript-login-example-signup.png
