+++ client/src/services/auth-header.ts
