+++ server/package.json
@@ -1,0 +1,31 @@
+{
+	"name": "server",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"build": "npx tsc",
+    "start": "node dist/index.js",
+    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\""
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"@types/bcrypt": "^5.0.0",
+		"@types/jsonwebtoken": "^9.0.1",
+		"@types/mongoose": "^5.11.97",
+		"cors": "^2.8.5",
+		"dotenv": "^16.0.3",
+		"express": "^4.18.2",
+		"morgan": "^1.10.0",
+		"ts-node": "^10.9.1"
+	},
+	"devDependencies": {
+		"@types/express": "^4.17.17",
+		"@types/node": "^18.15.3",
+		"concurrently": "^7.6.0",
+		"nodemon": "^2.0.21",
+		"typescript": "^4.9.5"
+	}
+}
